bump prod (#44)
* fix: crap ton of changes and bug fixes

* feat: forgot to add texture here

* fix: 64x32 skulls being white

* fix: back left being flipped

* fix: openapi docs

* fix: endpoint

* chore: regenerate openapi schema

* refactor: small consistency fixes, add consumable secton

* feat: sort items based on the inventory index as well, not only rarity, so it's like befroe

* fix: sunflower garden milestone

* feat: add sea creatures

* fix: replacing entire accessory data for dupes

* feat: add mutations

* feat: add attribute_shards endpoint

* perf: do not fetch if it's bot

* feat: add missing accessories

* perf: better username resolving caching

* feat: add ability to toggle verbose logging & forensics

* feat: add new museum categories

* Fix: Bukkit Case 351 damage 2 maps to cactus_green (#37)

* Fix: case 351 wrong dye mapping (cyan → cactus_green)

* Updated Cases 402 and 405

Some fucktwat tried at least lol

* perf: make startup a lot faster, especially dev

* Fix: use green_dye instead of cactus_green for dye ID 351 (#38)

* panic: failed to connect to Redis: could not connect to Redis: LOADING Redis is loading the dataset in memory

* refactor: do not use go git

* fix: start update loop only on main thread

* fix: update on all threads, forgot to add on last commit

* fix: trophy fish

* feat(inventory): add sacks

* very important fix, trust me nocap

* fix: error

* fix: hypixelCase at SkyCrypt, gen types

* fix: consumables texture

* fix: foraging lvl cap

* fix: embed

* build: bump mongodb to v2

* fix: FS link

* H+ link

* feat: add rank to the embed

* bump docs

* disable bot detection to test smth

* reenable bot detection

* change format of embed data

* fix: datatype

---------

+++ src/models/embed.go
@@ -30,7 +30,7 @@ type EmbedDataDungeons struct {
 }
 
 type EmbedDataSlayers struct {
-	Experience float64        `json:"xp"`
+	Experience int            `json:"xp"`
 	Slayers    map[string]int `json:"slayers"`
 }
